Set up MVCI framework for the GUI chatbox.
Following the Model-View-Controller-Interactor framework, added new classes for each component. The view layout follows the POC image from issue #24. The model reflects the properties the user can change in the GUI, which are set up by the controller. The interactor will have the logic for triggering inference and updating the output displays.