Implement interactive mode for GUI chatbox.
Created a new method `runInteractiveStep()` in Model.java for the GUI to use when interactive mode is selected. It uses a simple data object Model.Response, which keeps track of a chat session's conversation tokens, state, and TornadoVM plan while running. Every time the GUI runs this method, a new Response object is created with updated state, and ChatboxInteractor saves this response as part of the ongoing chat.